requirements.txt
streamlit
pandas
pytesseract
Pillow
fuzzywuzzy
pdf2image
fpdf
